# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "SMS MANAGEMENT API",
    "version": "1.0.0",
    "title": "SMS Management API",
    "contact": {
      "email": "ezrqnkemboi@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8000",
  "basePath": "/",
  "tags": [
    {
      "name": "Messages",
      "description": "Messages Management"
    },
    {
      "name": "Contacts",
      "description": "Contacts Management"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Homepage"
        ],
        "summary": "Homepage URL",
        "description": "Get homepage URL",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Welcome to SMS managament application."
          }
        }
      }
    },
    "/contacts": {
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Add new Contact",
        "description": "Adding new contacts",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "New CONTACT Object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Contact"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Contact Created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get All Contacts",
        "description": "Get all contacts in the system",
        "operationId": "getContacts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Contact"
              }
            }
          }
        }
      }
    },
    "/contacts/{id}": {
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete Contact",
        "description": "Delete Single Contact with all messages related to it",
        "operationId": "deleteContact",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of the contact to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted contacts and sms related to them successfully"
          },
          "404": {
            "description": "Invalid id supplied"
          }
        }
      }
    },
    "/sms": {
      "post": {
        "tags": [
          "Sms"
        ],
        "summary": "Send new sms message",
        "description": "Add new sms to the DB",
        "operationId": "sendNewMessage",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "New SMS Object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SMS"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message was sent successfully"
          }
        }
      },
      "get": {
        "tags": [
          "Sms"
        ],
        "summary": "Get All sms Messages",
        "description": "Get all messages regardless of the sender",
        "operationId": "getAll",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "All sms retrieved successfully",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SMS"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "SMS": {
      "type": "object",
      "properties": {
        "senderId": {
          "type": "integer"
        },
        "receiverId": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "Contact": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      }
    }
  }
}